requirements.txt
Django==1.6.1
django-mease==0.1.7
djangorestframework==2.3.12
